requirements.txt
asgiref==3.2.10
Django==3.1.2
django-rest-framework==0.1.0
djangorestframework==3.12.1
psycopg2-binary==2.8.6
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
view==0.1
whitenoise==5.2.0
gunicorn==20.0.4
